refactor: move source map functionality from agent to mapping-file
This is to not include this code in the browser recorder.
