--- components/agent/default/.ordering
@@ -1,2 +1,0 @@
-source-map.mjs
-index.mjs
