+++ components/mapping-file/node/.env
@@ -1,0 +1,2 @@
+test
+node
